Refactor provider module
